feat: Add browser tools for session management and DOM interaction
- Implemented ClickElementTool for clicking on DOM elements using CSS selectors.
- Created CloseSessionTool to close existing browser sessions.
- Added CreateSessionTool to initiate new browser sessions.
- Developed ExtractContentTool to extract page content as Markdown or JSON.
- Introduced FindElementTool to check the state of DOM elements.
- Implemented GetCurrentUrlTool to retrieve the current URL of the browser page.
- Created GetElementAttributeTool to get specific attribute values from elements.
- Added GetElementTextTool to retrieve text content from elements.
- Developed GetPageTitleTool to fetch the title of the current page.
- Implemented InputTextTool for inputting text into form fields.
- Created ListSessionsTool to list all active browser sessions.
- Added NavigateBackTool and NavigateForwardTool for browser history navigation.
- Implemented NavigateToUrlTool to navigate to a specified URL.
- Developed ScrollPageTool to scroll the page to specified coordinates.
- Introduced helper functions and schemas for browser tools in helpers.ts.
- Defined types for local and browser tools in types.ts.
- Updated index.ts to export all new tools and types.
