+++ refactoring_plan_submission_guide.md
@@ -7,6 +7,7 @@
 - 추가 분석 과제 (선택적)
 - 변경 이후의 상태 / 해결 판정 기준
 - 수정이 필요한 코드 및 수정부분의 코드 스니핏
+- 재사용 가능한 연관 코드 (파일 경로, 주요 기능, 인터페이스 등 포함)
 - file 규칙
   - ./docs/history/refactoring_{yyyyMMdd_hhmm}.md
