+++ src/features/tools/browser-tools/index.ts
@@ -1,0 +1,29 @@
+// Simple tools (no executeScript dependency)
+export { createSessionTool } from './CreateSessionTool';
+export { closeSessionTool } from './CloseSessionTool';
+export { listSessionsTool } from './ListSessionsTool';
+export { navigateToUrlTool } from './NavigateToUrlTool';
+
+// Tools requiring executeScript
+export { getCurrentUrlTool } from './GetCurrentUrlTool';
+export { getPageTitleTool } from './GetPageTitleTool';
+export { scrollPageTool } from './ScrollPageTool';
+export { navigateBackTool } from './NavigateBackTool';
+export { navigateForwardTool } from './NavigateForwardTool';
+export { getElementTextTool } from './GetElementTextTool';
+export { getElementAttributeTool } from './GetElementAttributeTool';
+export { findElementTool } from './FindElementTool';
+
+// Complex tools
+export { clickElementTool } from './ClickElementTool';
+export { inputTextTool } from './InputTextTool';
+export { extractContentTool } from './ExtractContentTool';
+
+// Types and helpers
+export type { LocalMCPTool, BrowserLocalMCPTool } from './types';
+export {
+  BROWSER_TOOL_SCHEMAS,
+  checkElementState,
+  pollWithTimeout,
+  formatBrowserResult,
+} from './helpers';
